mise à jour de la fonction de séparation, mise en place d'un slider pour changer le rebond quand les boids arrivent dans la protection range

+++ include/interface.hpp
@@ -18,7 +18,7 @@ private: // on déclare les variables membres de la classe Interface
     glm::vec2    position_cercle;
     int          nombre_boids;
     float        taille_boids;
-    int          separation;
+    float        separation;
     int          alignement;
     int          cohesion;
     std::string  texte;
